backend/requirements.txt
google-generativeai
PyGithub
Flask
python-dotenv
flask-cors